Provide rx queues for convenience
NOTE: This is a breaking change.

Previously, callers of rx_char and rx_keyboard would have to check
output to see if the UART was actually ready to receive a character.
This change adds a receive queue to the DUART implementation that will
accept characters regardless of whether the DUART is ready to receive
or not. The service routine will then read fromn the queue if and only
if the DUART is ready to transmit.